infinte loop fixed

+++ index.js
@@ -1,14 +1,19 @@
 const Discord = require('discord.js');
-const { prefix, token } = require('./config.json'); //prefix is the sympol you right before the command for example !play in rythm I can't show you the file as it's contain my token :D
+const {
+    prefix,
+    token
+} = require('./config.json'); //prefix is the sympol you right before the command for example !play in rythm I can't show you the file as it's contain my token :D
 const client = new Discord.Client();
-const { restrictedWords } = require('./wordList.json');
+const {
+    restrictedWords
+} = require('./wordList.json');
 
 const http = require('http');
 
 http.createServer((req, res) => {
-res.writeHead(200, {
-    'Content-type': 'text/plain'
-});
+    res.writeHead(200, {
+        'Content-type': 'text/plain'
+    });
     res.write('Hey');
     res.end();
 }).listen(4000);
@@ -19,43 +24,49 @@ client.once('ready', () => {
 
 
 client.on('message', message => {
+
     // console.log(message.content);
-    let member = message.mentions.members.first();
 
-    if (message.content.startsWith(`${prefix}sayHello`)) {
-        message.delete();
-        message.channel.send("Hello " + member + ":wave:");
-    }
-    
-    if (message.content.startsWith(`${prefix}rankMe`)) {
+    if (!message.author.bot) {
+
         let member = message.mentions.members.first();
-        message.channel.send('checking data-base...')
-            .then(() => {
-                message.channel.send(`${member.displayName} is confirmed to be boosted`)
+
+        if (message.content.startsWith(`${prefix}sayHello`)) {
+            message.delete();
+            message.channel.send("Hello " + member + ":wave:");
+        }
+
+        if (message.content.startsWith(`${prefix}rankMe`)) {
+            let member = message.mentions.members.first();
+            message.channel.send('checking data-base...')
+                .then(() => {
+                    message.channel.send(`${member.displayName} is confirmed to be boosted`)
+                });
+        }
+
+        if (message.content.startsWith(`${prefix}fuck`)) {
+            let member = message.author.username;
+            message.channel.send(`Ah yes ${member} do you want more ?`).then(message => {
+                setTimeout(() => {
+                    message.edit(`${member} is confirmed to be gay`)
+                }, 2000) // a3ml save
             });
+        }
+
+        if (message.content.match(/Do u love me/i) ||
+            message.content.match(/Do u want me/i) ||
+            message.content.match(/Do u need me/i)) {
+            message.channel.send('Do u do u');
+        }
+
+
+        if (message.content.match(/Sasageyo/i)) {
+            message.channel.send('Shinzou wo Sasageyo');
+        }
+
+
     }
 
-    if (message.content.startsWith(`${prefix}fuck`)) {
-        let member = message.author.username;
-        message.channel.send(`Ah yes ${member} do you want more ?`).then(message => {
-            setTimeout(() => {
-                message.edit(`${member} is confirmed to be gay`)
-            }, 2000) // a3ml save
-        });
-    }
-
-    if (message.content.match(/Sasageyo/i)) {
-        message.channel.send("Shinzou wo Sasageyo");
-    }
-
-    if (message.content.match(/Do u love me/i) || 
-        message.content.match(/Do u want me/i) || 
-        message.content.match(/Do u need me/i)) {
-        message.channel.send('Do u do u');
-    }
-
-    
-    
     // const exampleEmbed = new Discord.RichEmbed().setTitle('Warning');
     // if (!message.author.bot)  {
     //     restrictedWords.map(word => {
@@ -84,7 +95,7 @@ client.on('message', message => {
     // }
     // checks if the message from the bot itself
     // } filter the messages
-    
+
     if (message.content.includes('بلحة')) {
         message.channel.send("اسمه الرئيس المشير عبدالفتاح السيسي يا عدو الوطن ");
     }
